requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
imap-tools==1.5.0
chromadb==0.4.15
ollama==0.1.6
pydantic==2.5.0
beautifulsoup4==4.12.2
lxml==4.9.3